skills/nano-website-style-images/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
PyYAML>=6.0
